+++ 04-vue-cli/02-TheToaster/components/UiToaster.vue
@@ -1,0 +1,31 @@
+<template>
+  <div class="toasts">
+    <slot />
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'UiToaster',
+  };
+</script>
+
+<style scoped>
+  .toasts {
+    position: fixed;
+    bottom: 8px;
+    right: 8px;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    white-space: pre-wrap;
+    z-index: 999;
+  }
+
+  @media all and (min-width: 992px) {
+    .toasts {
+      bottom: 72px;
+      right: 112px;
+    }
+  }
+</style>
