--- 04-vue-cli/02-TheToaster/components/UiToastsContainer.vue
@@ -1,70 +1,0 @@
-<template>
-  <div class="toasts">
-    <ui-toast v-for="toast in toasts" :key="toast.id" v-bind="toast"/>
-  </div>
-</template>
-
-<script>
-import UiToast from './UiToast';
-
-export default {
-  name: 'UiToastsContainer',
-
-  components: { UiToast },
-
-  provide() {
-    return {
-      toaster: {
-        removeToast: this.removeToast,
-      }
-    }
-  },
-
-  data() {
-    return {
-      toasts: [],
-    }
-  },
-
-  props: {
-    toastTimeout: {type: Number, default: 5000},
-    needCloseButton: {type: Boolean, default: false},
-  },
-
-  methods: {
-    addToast({type='error', message}) {
-      const toastObj = Object.assign(
-        {id: (new Date()).valueOf(), close: this.needCloseButton},
-        arguments[0]
-      )
-      this.toasts.push(toastObj)
-      setTimeout(()=>{this.removeToast(toastObj.id)}, this.toastTimeout)
-    },
-
-    removeToast(toastId) {
-      this.toasts.splice(this.toasts.findIndex(x=>x.id===toastId), 1)
-    },
-  },
-};
-</script>
-
-<style scoped>
-.toasts {
-  position: fixed;
-  bottom: 8px;
-  right: 8px;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  white-space: pre-wrap;
-  z-index: 999;
-}
-
-@media all and (min-width: 992px) {
-  .toasts {
-    bottom: 72px;
-    right: 112px;
-  }
-}
-
-</style>
